backend/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
click==8.1.8
colorama==0.4.6
distlib==0.3.9
fastapi==0.115.6
filelock==3.16.1
h11==0.14.0
idna==3.10
platformdirs==4.3.6
pydantic==2.10.5
pydantic_core==2.27.2
python-multipart==0.0.20
sniffio==1.3.1
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.34.0
virtualenv==20.27.1
SQLAlchemy==2.0.25
psycopg2-binary==2.9.9
python-dotenv==1.0.0
